+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Changes
 
--   Add Interventions and coordinates tabs to employee details
+-   Add Interventions and contact details tabs to employee details
 
 # [0.8.3] (2025-09-03)